Update to Version 1.01.20200807a
 - Removed a commented section of code.
 - I added a few fields to skip under the 'claims' section of a character's information.

Small bug: See comment in code.  Under the claims= section, the additionally indentend '\ttitle=' field appears to be duplicated on custom/dynamic titles.  Will investigate, but it's only a small hiccup that I can correct as needed on the with some WD40 and duct tape.  I'm uncertain if this bug also carries over to the standard (non-dynamic/non-custom) titles.

New files: I'm incorporating logic into the character statistic shifter for religions.  It's mostly if I get a wild hair and want to see if I put a religion in a geographic location that would have made no sense in the Middle Ages.